indexing/requirements.txt
# RAG indexing pipeline
azure-search-documents>=11.4.0
azure-identity>=1.15.0
openai>=1.12.0
python-dotenv>=1.0.0
